feat: context edit profile modal
