+++ src/components/Modal/EditProfile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./EditProfileModal";
